# docs/openapi/openapi.bundle.yaml
openapi: 3.0.3
info:
  title: rosmon HTTP API
  version: 1.0.0
  description: |
    REST + Server-Sent Events service di atas RouterOS API. Multi-router:
    semua endpoint (kecuali Device CRUD & Health) berada di bawah
    `/api/v1/devices/{device_id}/...` agar satu instance bisa melayani
    banyak router sekaligus.

    ## Response envelope

    Semua endpoint pakai envelope konsisten:

    - **Success single**: `{ "data": { ... } }`
    - **Success list**:   `{ "data": [ ... ], "meta": { "count": N } }`
    - **Error**:          `{ "error": { "code": "...", "message": "...", "path": "..." } }`

    Error `code` stabil — frontend bisa branch dengan aman. Mapping:

    | Code | HTTP | Trigger |
    |---|---|---|
    | `NOT_FOUND` | 404 | `mikrotik.ErrNotFound` |
    | `INVALID_ARGUMENT` | 400 | `mikrotik.ErrInvalidArgument` |
    | `AMBIGUOUS` | 409 | `mikrotik.ErrAmbiguous` |
    | `VALIDATION` | 400 | body gagal validasi `gin` binding |
    | `TIMEOUT` | 504 | `context.DeadlineExceeded` |
    | `CANCELED` | 499 | `context.Canceled` |
    | `INTERNAL` | 500 | uncategorized / panic recovery |
    | `DEVICE_NOT_FOUND` | 404 | device middleware: device ID tidak terhubung |
    | `SERVICE_UNAVAILABLE` | 503 | InfluxDB / dependency lain down |

    Setiap response menyertakan header `X-Request-ID` (16-char hex) untuk
    korelasi log.

    ## Authentication

    Service ini sekarang mendukung **JWT access + refresh token**
...

...
    4. `POST /auth/logout` revoke refresh token.

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /hook/hotspot/login/{device_id}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
    post:
      tags:
        - Webhook
      summary: Webhook on-login dari MikroTik (catat selling record)
      security: []
      description: |
        **Endpoint internal — dipanggil oleh on-login script MikroTik via
        `/tool/fetch`.** Tidak butuh JWT (router tidak bisa pegang token).

        **PERINGATAN MVP**: tidak ada IP whitelist atau shared secret.
        Jangan expose port ke internet publik. Pakai firewall rule supaya
        hanya reachable dari MikroTik router.

        Flow:

          1. Validasi `device_id` exists di DB.
          2. Lookup `profile_config (device_id, profile)`.
          3. Skip kalau mode tidak mencatat transaksi — hanya `remc` dan `ntfc`
             yang lanjut. Mode `rem`, `ntf`, dan `"0"` (free) menghasilkan
             `reason: non_recording_mode`.
          4. Dedup berbasis window validity: skip kalau sudah ada record
             `(device_id, username, comment="login")` yang dibuat dalam window
             `validity` terakhir (re-login = koneksi ulang, bukan pembelian baru).
             Login setelah window habis ter-record sebagai pembelian baru.
          5. Insert transaction baru dengan price/sell_price/validity dari
             profile_config.

        Selalu return 200 OK untuk error non-fatal (router tidak punya retry
        yang bermakna). Body response berisi `recorded` (bool) + `reason`
        (skip reason kalau ada).
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - user
              properties:
                user:
                  type: string
                  description: Hotspot username (dari `$user`).
                  example: vc-12345
                  maxLength: 128
                mac:
                  type: string
                  description: Client MAC address (dari `$"mac-address"`).
                  example: AA:BB:CC:DD:EE:FF
                  maxLength: 17
                ip:
                  type: string
                  description: Client IP (dari `$address`).
                  example: 10.0.0.5
                  maxLength: 45
                profile:
                  type: string
                  description: |
                    Profile name (di-embed di script saat generate, tidak
                    perlu lookup ke router). Kalau kosong → skip recording.
                  example: 1day
                  maxLength: 64
      responses:
        '200':
          description: Webhook processed (recorded atau di-skip)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          recorded:
                            type: boolean
                            description: True kalau transaction baru di-insert.
                          reason:
                            type: string
                            description: |
                              Diisi kalau `recorded=false`. Nilai:
                              `missing_profile`, `profile_not_configured`,
                              `non_recording_mode`, `already_recorded`.
                            enum:
                              - missing_profile
                              - profile_not_configured
                              - non_recording_mode
                              - already_recorded
                          transaction_id:
                            type: integer
                            description: Diisi kalau `recorded=true`.
        '400':
          description: Bad request (invalid device id atau missing user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Device tidak ditemukan di DB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error (DB write failed, dst)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /devices/{device_id}/hotspot/active:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/ServerQuery'
    get:
      tags:
        - Hotspot Sessions
      summary: List active sessions
      description: List session hotspot yang sedang aktif (sudah login). Filter `?server=`.
      responses:
        '200':
          description: Active session list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HotspotActiveResponse'
  /devices/{device_id}/hotspot/active/count:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/ServerQuery'
    get:
      tags:
        - Hotspot Sessions
      summary: Count active sessions
      responses:
        '200':
          description: Count
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CountResponse'
  /devices/{device_id}/hotspot/active/{id}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/IdPath'
    get:
      tags:
        - Hotspot Sessions
      summary: Get active session by ID
      responses:
        '200':
          description: Session detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HotspotActiveResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    delete:
      tags:
        - Hotspot Sessions
      summary: Kick (disconnect) session
      description: |
        Workflow KickActive: hapus cookie user (anti auto-relogin) → hapus
        active record. Tidak hapus user dari `/hotspot/user`.
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
  /devices/{device_id}/ppp/active:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
    get:
      tags:
        - PPP
      summary: List active PPP sessions
      responses:
        '200':
          description: Active session list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PPPActiveResponse'
  /devices/{device_id}/ppp/active/{id}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/IdPath'
    delete:
      tags:
        - PPP
      summary: Disconnect PPP session
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /devices/{device_id}/stream/hotspot/active:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/ModeQuery'
    get:
      tags:
        - Streaming
      summary: Stream hotspot active sessions (SSE)
      description: |
        Server-Sent Events stream untuk session hotspot.

        **Format event**:
        ```
        event: change
        id: <event-id>
        data: <HotspotActiveEvent JSON>

        ```

        **Default mode**: snapshot+follow. Subscribe akan kirim snapshot row
        yang aktif sekarang, lalu follow delta. Pakai `?mode=follow-only`
        untuk skip snapshot.

        Broker auto-stop saat last subscriber leave; multi-subscriber per
        topic share single backend stream ke router.

        **Sentinel field**: `dead: true` artinya session terminate.
        ID-nya tetap valid untuk cross-ref snapshot.
      x-event-types:
        - change
      responses:
        '200':
          description: SSE stream (Content-Type text/event-stream)
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/HotspotActiveEvent'
        '404':
          $ref: '#/components/responses/ErrorDeviceNotFound'
...
  /devices/{device_id}/stream/ppp/active:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
    get:
      tags:
        - Streaming
      summary: Stream PPP active sessions (SSE)
      description: |
        SSE delta sesi PPP. Format identik dengan hotspot active:
        `event: change` + `data: <PPPActiveEvent JSON>`. Sentinel
        `dead: true` saat session terminate.
      x-event-types:
        - change
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/PPPActiveEvent'
        '404':
          $ref: '#/components/responses/ErrorDeviceNotFound'
...
  /devices/{device_id}/history/hotspot/active:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/FromQuery'
      - $ref: '#/components/parameters/ToQuery'
    get:
      tags:
        - History
      summary: Query historical hotspot connect/disconnect events
      description: 'Kolom: `time, user, event ("connect"|"disconnect"), session_id, mac, address`.'
      responses:
        '200':
          description: History rows
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HistoryRow'
        '503':
          $ref: '#/components/responses/ErrorServiceUnavailable'
  /devices/{device_id}/history/ppp/active:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/FromQuery'
      - $ref: '#/components/parameters/ToQuery'
    get:
      tags:
        - History
      summary: Query historical PPP connect/disconnect events
      description: 'Kolom: `time, user, event, address`.'
      responses:
        '200':
          description: History rows
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HistoryRow'
        '503':
          $ref: '#/components/responses/ErrorServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (HS256). Diperoleh via `POST /auth/login`.
        Wajib di-include sebagai header `Authorization: Bearer <token>`
        kecuali untuk endpoint public (login, refresh, logout, /healthz).
  schemas:
    Envelope:
      type: object
      description: Response envelope generik. Field `data` dan `error` saling eksklusif.
      properties:
        data:
          description: Payload sukses. Tipe spesifik tergantung endpoint.
        meta:
          type: object
          additionalProperties: true
          description: Metadata opsional, biasanya berisi `count` untuk list.
          example:
            count: 47
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorEnvelope:
      type: object
      description: Envelope untuk response error (status 4xx/5xx).
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorPayload:
      type: object
      description: Detail error yang stabil — frontend bisa branch via `code`.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable error code untuk branching di client.
          enum:
            - NOT_FOUND
            - INVALID_ARGUMENT
            - AMBIGUOUS
            - VALIDATION
            - TIMEOUT
            - CANCELED
            - INTERNAL
            - DEVICE_NOT_FOUND
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Pesan error human-readable.
          example: user with id *99 not found
        path:
          type: string
          description: Path request yang menghasilkan error.
          example: /api/v1/devices/rb-main/hotspot/users/*99
        details:
          description: Payload tambahan (mis. validator detail). Tipe bebas.
    CountResponse:
      type: object
      description: Body untuk endpoint count-only.
      required:
        - count
      properties:
        count:
          type: integer
          format: int64
          minimum: 0
          example: 142
...
    HotspotActiveResponse:
      type: object
      description: Sesi hotspot yang sedang aktif (sudah login).
      required:
        - id
        - user
        - disabled
      properties:
        id:
          type: string
          example: '*1'
        user:
          type: string
          example: user01
        address:
          type: string
          example: 10.0.0.5
        mac_address:
          type: string
          example: AA:BB:CC:DD:EE:FF
        server:
          type: string
          example: hotspot1
        login_by:
          type: string
          example: http-pap
        uptime:
          type: string
          example: 12m34s
        bytes_in:
          type: integer
          format: int64
        bytes_out:
          type: integer
          format: int64
        packets_in:
          type: integer
          format: int64
        packets_out:
          type: integer
          format: int64
        idle_time:
          type: string
          example: 5s
        session_time_left:
          type: string
          example: 47m
        keepalive_timeout:
          type: string
        comment:
          type: string
...
    PPPActiveResponse:
      type: object
      description: |
        Session PPP aktif. Field paling penting untuk UI monitoring:
        uptime, address, caller_id.
      required:
        - id
        - name
      properties:
        id:
          type: string
          example: '*1'
        name:
          type: string
          example: pppoe-user01
        service:
          type: string
          example: pppoe
        caller_id:
          type: string
          example: AA:BB:CC:DD:EE:FF
        address:
          type: string
          example: 10.0.0.10
        uptime:
          type: string
          example: 1h23m
        limit_bytes_in:
          type: integer
          format: int64
        limit_bytes_out:
          type: integer
          format: int64
        comment:
          type: string
    HotspotActiveEvent:
      type: object
      description: |
        Event delta dari /stream/hotspot/active. Kalau `dead=true`, session
        telah terminate; ID-nya valid untuk cross-ref ke snapshot terakhir
        tapi field lain mungkin kosong.
      required:
        - id
        - user
        - bytes_in
        - bytes_out
        - dead
      properties:
        id:
          type: string
          example: '*1'
        user:
          type: string
          example: user01
        address:
          type: string
          example: 10.0.0.5
        mac_address:
          type: string
          example: AA:BB:CC:DD:EE:FF
        server:
          type: string
          example: hotspot1
        login_by:
          type: string
          example: http-pap
        bytes_in:
          type: integer
          format: int64
        bytes_out:
          type: integer
          format: int64
        uptime:
          type: string
          example: 12s
        dead:
          type: boolean
          description: true = session terminated.
    PPPActiveEvent:
      type: object
      required:
        - id
        - name
        - dead
      properties:
        id:
          type: string
          example: '*1'
        name:
          type: string
          example: pppoe-user01
        service:
          type: string
        caller_id:
          type: string
        address:
          type: string
        uptime:
          type: string
        dead:
          type: boolean
          description: true = session terminated.
...
    HistoryRow:
      type: object
      description: |
        Satu baris hasil query InfluxDB. Kolom bervariasi per endpoint:

        - `/history/resource`: `time, cpu_load, free_memory, free_hdd_space`.
        - `/history/interfaces`: `time, interface, rx_bytes, tx_bytes, rx_pps, tx_pps`.
        - `/history/hotspot/users`: `time, user, profile, bytes_in_delta, bytes_out_delta`.
        - `/history/hotspot/active`: `time, user, event ("connect"|"disconnect"), session_id, mac, address`.
        - `/history/ppp/active`: `time, user, event, address`.
        - `/history/queues`: `time, queue, max_limit, bytes_in_delta, bytes_out_delta`.

        Skema dinamis (`additionalProperties: true`) supaya tidak berubah saat
        kolom InfluxDB di-extend. Value bisa scalar (string/number/bool) atau
        timestamp (RFC3339).
      additionalProperties: true
      example:
        time: '2026-05-18T08:00:00Z'
        cpu_load: 12
        free_memory: 67108864
        free_hdd_space: 100663296
...
  parameters:
    DeviceIdPath:
      name: device_id
      in: path
      required: true
      description: Numeric device ID.
      schema:
        type: integer
        format: int64
        example: 1
    IdPath:
      name: id
      in: path
      required: true
      description: RouterOS internal ID (`.id`), biasanya format `*N` (mis. `*1`, `*2A`).
      schema:
        type: string
        example: '*1'
...
    FromQuery:
      name: from
      in: query
      required: false
      description: Awal range waktu (RFC3339). Default `now-1h`.
      schema:
        type: string
        format: date-time
        example: '2026-05-18T08:00:00Z'
    ToQuery:
      name: to
      in: query
      required: false
      description: Akhir range waktu (RFC3339). Default `now`.
      schema:
        type: string
        format: date-time
        example: '2026-05-18T09:00:00Z'
...
    ServerQuery:
      name: server
      in: query
      required: false
      description: Filter nama hotspot server.
      schema:
        type: string
        example: hotspot1
...
    ModeQuery:
      name: mode
      in: query
      required: false
      description: Mode streaming. `follow-only` = tanpa snapshot awal (event baru saja).
      schema:
        type: string
        enum:
          - follow-only
...
  responses:
...
    ErrorNotFound:
      description: Resource yang dirujuk tidak ditemukan.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            user_not_found:
              summary: User tidak ada
              value:
                error:
                  code: NOT_FOUND
                  message: 'mikrotik: not found'
                  path: /api/v1/devices/rb-main/hotspot/users/*99
    ErrorDeviceNotFound:
      description: Device tidak ditemukan atau tidak terhubung. Dikembalikan oleh DeviceMiddleware.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            device_offline:
              summary: Device offline
              value:
                error:
                  code: DEVICE_NOT_FOUND
                  message: device not found or not connected
                  path: /api/v1/devices/rb-main/system/identity
...
    ErrorServiceUnavailable:
      description: Dependency external (InfluxDB / DB) sedang tidak available.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            influx_down:
              summary: InfluxDB tidak terhubung
              value:
                error:
                  code: SERVICE_UNAVAILABLE
                  message: history backend unavailable
                  path: /api/v1/devices/rb-main/history/resource
...
    NoContentResponse:
      description: Operasi sukses tanpa body response (204 No Content).